Fix a bundling problem with Webpack
When attempting to bundle blurhash with Webpack, it throw this error and the bundling fails:

```
Module not found: Error: Default condition should be last one
```

This is because Webpack expects the default condition to be the last one in the exports object.

This problem was introduced in the NPM 2.0.1 release of blurhash.

See: https://webpack.js.org/guides/package-exports/#conditional-syntax